+++ swiyu-trust-management-service/src/main/java/ch/admin/bj/swiyu/trust/management/modules/management/domain/details/VerificationQueryV2Details.java
@@ -48,7 +48,9 @@ public final class VerificationQueryV2Details extends TrustStatementDetails {
     @AllArgsConstructor
     public enum Language {
         DEFAULT(""),
+        @Deprecated(forRemoval = true, since = "3.29.1") // Remove in EID-6303
         EN("en"),
+        EN_CH("en-CH"),
         DE_CH("de-CH"),
         FR_CH("fr-CH"),
         IT_CH("it-CH"),