+++ swiyu-trust-management-service/src/main/java/ch/admin/bj/swiyu/trust/management/modules/ui/api/TrustOnboardingTaskDto.java
@@ -36,7 +36,9 @@ public record TrustOnboardingTaskDto(
     @RequiredArgsConstructor
     @Schema(name = "Language", enumAsRef = true)
     public enum LanguageDto {
+        @Deprecated(forRemoval = true, since = "3.29.1") // Remove in EID-6303
         EN("en"),
+        EN_CH("en-CH"),
         DE_CH("de-CH"),
         FR_CH("fr-CH"),
         IT_CH("it-CH"),